refactor: move table definition to its own file
