+++ src/codegen/cloudformation/types.ts
@@ -1,0 +1,5 @@
+export interface CloudFormationFragment {
+  env?: Record<string, {Ref: string}>;
+  output?: Record<string, object>;
+  resources?: Record<string, object>;
+}
